chore: Update app.js to handle uncaught exceptions and unhandled promise rejections
This commit adds error handling code to the `app.js` file to handle uncaught exceptions and unhandled promise rejections. It uses the `process.on` method to listen for these events and logs the error message before shutting down the server. Additionally, it updates the `app.js` file to use the `errorMiddleware` for handling errors. This improves the robustness and stability of the application.It also handle Validation and Mongoose ID Error.