chore: Associate user with product in newProduct controller
This commit modifies the `newProduct` controller in the `productController.js` file to associate the user with the product being created. The user's ID is added to the `req.body.user` property before creating the product. This enhancement ensures that the product is linked to the correct user in the application.