feat: Add password reset functionality
This commit adds a new method `getResetPasswordToken` to the `userSchema` in the `user.js` file. This method generates a password reset token using the `crypto` module and sets the `resetPasswordToken` field in the user document. It also sets the `resetPasswordExpire` field to indicate the expiration time of the token. This enhancement allows users to request a password reset and receive a unique token for verification.