feat: Add getProductReviews endpoint
This commit adds the `getProductReviews` endpoint to the `productController.js` file. It allows users to retrieve the reviews for a specific product by providing the product ID in the request query. If the product is not found, an error with status code 404 is returned. This enhancement improves the application by providing users with the ability to view the reviews for a product.