feat: Improve user dropdown menu in Header component
This commit enhances the user dropdown menu in the Header component. It adds the ability to display the user's avatar, name, and options for navigating to the dashboard, orders, and profile pages. It also includes a logout option for users to securely log out of the application. This improvement improves the user experience by providing a more personalized and convenient way for users to access their account-related features.

This commit also  removes the unnecessary scripts for Popper.js and Bootstrap from the index.html file. These scripts were not being used in the application and were causing unnecessary network requests.