feat: Update npm dependencies
This commit updates the npm dependencies in the package.json and package-lock.json files. It includes updating the bcryptjs dependency to version 2.4.3 and adding the cloudinary dependency with version 2.2.0. This update ensures that the project is using the latest stable versions of the dependencies and introduces the cloudinary library for handling avatar uploads.