feat: Update cart functionality to remove items from the cart
This commit updates the cart functionality in the frontend codebase to allow users to remove items from their shopping cart. The `removeCartItem` reducer function is added to the `cartSlice.js` file, which filters out the item to be removed from the `cartItem` state. Additionally, a toast notification is displayed to inform the user that the item has been successfully removed. This improvement enhances the user experience by providing a convenient way for users to manage their shopping cart and remove unwanted items.